refactor(neovim-plugins): change event triggers for various plugins
This commit modifies the event triggers for several Neovim plugins. The 'BufEnter' event has been replaced with 'InsertEnter' for nvim-autopairs and nvim-dap plugins. The 'VeryLazy' event has been added to neodev.nvim and nvim-lspconfig plugins. This change optimizes the plugin loading process and improves the overall performance of the Neovim setup.